[mobile]refactor: extract HomeScreen
