+++ mobile/app/src/main/java/com/thoughtworks/mobile/ui/fragments/home/HomeFragment.kt
@@ -5,29 +5,11 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
-import androidx.compose.ui.unit.dp
-import androidx.lifecycle.compose.ExperimentalLifecycleComposeApi
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.fragment.findNavController
 import com.thoughtworks.mobile.R
-import com.thoughtworks.mobile.ui.fragments.HomeUiState
 import com.thoughtworks.mobile.ui.fragments.HomeViewModel
-import com.thoughtworks.mobile.ui.theme.MobileTheme
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class HomeFragment : Fragment() {
@@ -45,36 +27,6 @@ class HomeFragment : Fragment() {
                     findNavController().navigate(R.id.action_home_to_add)
                 }
             }
-        }
-    }
-}
-
-@OptIn(ExperimentalLifecycleComposeApi::class)
-@Composable
-fun HomeScreen(viewModel: HomeViewModel, onClick: () -> Unit) {
-    val uiState: HomeUiState by viewModel.uiState.collectAsStateWithLifecycle()
-
-    MobileTheme {
-        Surface(
-            modifier = Modifier.fillMaxSize(),
-            color = MaterialTheme.colorScheme.background
-        ) {
-            FloatingButton(onClick = onClick)
-        }
-    }
-}
-
-@Composable
-fun FloatingButton(onClick: () -> Unit) {
-    Box(
-        modifier = Modifier
-            .padding(16.dp),
-        contentAlignment = Alignment.BottomEnd
-    ) {
-        FloatingActionButton(
-            onClick = { onClick() },
-        ) {
-            Icon(Icons.Filled.Add, "Add book button")
         }
     }
 }
