+++ mobile/app/src/main/res/navigation/nav_graph_home.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/home"
-        android:name="com.thoughtworks.mobile.ui.fragments.HomeFragment"
+        android:name="com.thoughtworks.mobile.ui.fragments.home.HomeFragment"
         android:label="HomeFragment">
         <action
             android:id="@+id/action_home_to_add"
